Moved studentDb to pandautils.
