+++ apps/student/src/main/java/com/instructure/student/di/DatabaseModule.kt
@@ -6,12 +6,12 @@ import com.instructure.pandautils.room.appdatabase.AppDatabase
 import com.instructure.pandautils.room.appdatabase.appDatabaseMigrations
 import com.instructure.pandautils.room.calendar.CalendarFilterDatabase
 import com.instructure.pandautils.room.calendar.calendarDatabaseMigrations
-import com.instructure.student.room.StudentDb
-import com.instructure.student.room.entities.daos.CreateFileSubmissionDao
-import com.instructure.student.room.entities.daos.CreatePendingSubmissionCommentDao
-import com.instructure.student.room.entities.daos.CreateSubmissionCommentFileDao
-import com.instructure.student.room.entities.daos.CreateSubmissionDao
-import com.instructure.student.room.studentDbMigrations
+import com.instructure.pandautils.room.studentdb.StudentDb
+import com.instructure.pandautils.room.studentdb.entities.daos.CreateFileSubmissionDao
+import com.instructure.pandautils.room.studentdb.entities.daos.CreatePendingSubmissionCommentDao
+import com.instructure.pandautils.room.studentdb.entities.daos.CreateSubmissionCommentFileDao
+import com.instructure.pandautils.room.studentdb.entities.daos.CreateSubmissionDao
+import com.instructure.pandautils.room.studentdb.studentDbMigrations
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
