+++ apps/student/src/main/java/com/instructure/student/mobius/assignmentDetails/submission/file/UploadStatusSubmissionEffectHandler.kt
@@ -20,8 +20,8 @@ import com.instructure.canvasapi2.utils.exhaustive
 import com.instructure.student.mobius.assignmentDetails.submission.file.ui.UploadStatusSubmissionView
 import com.instructure.student.mobius.common.ui.EffectHandler
 import com.instructure.student.mobius.common.ui.SubmissionHelper
-import com.instructure.student.room.StudentDb
-import com.instructure.student.room.entities.CreateFileSubmissionEntity
+import com.instructure.pandautils.room.studentdb.StudentDb
+import com.instructure.pandautils.room.studentdb.entities.CreateFileSubmissionEntity
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
