+++ apps/student/src/test/java/com/instructure/student/features/assignments/details/StudentAssignmentDetailsSubmissionHandlerTest.kt
@@ -26,8 +26,8 @@ import com.instructure.pandautils.features.assignments.details.AssignmentDetails
 import com.instructure.pandautils.utils.ColorKeeper
 import com.instructure.pandautils.utils.ThemePrefs
 import com.instructure.student.mobius.common.ui.SubmissionHelper
-import com.instructure.student.room.StudentDb
-import com.instructure.student.room.entities.CreateSubmissionEntity
+import com.instructure.pandautils.room.studentdb.StudentDb
+import com.instructure.pandautils.room.studentdb.entities.CreateSubmissionEntity
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.mockkStatic
