+++ libs/pandautils/src/main/java/com/instructure/pandautils/room/studentdb/StudentDbMigration.kt
@@ -12,7 +12,7 @@
  *     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  *     See the License for the specific language governing permissions and
  *     limitations under the License.
- */package com.instructure.student.room
+ */package com.instructure.pandautils.room.studentdb
 
 import com.instructure.pandautils.room.common.createMigration
