+++ libs/pandautils/src/main/java/com/instructure/pandautils/room/studentdb/entities/CreatePendingSubmissionCommentEntity.kt
@@ -13,7 +13,7 @@
  *     See the License for the specific language governing permissions and
  *     limitations under the License.
  */
-package com.instructure.student.room.entities
+package com.instructure.pandautils.room.studentdb.entities
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
