+++ libs/pandautils/src/main/java/com/instructure/pandautils/room/studentdb/entities/daos/CreatePendingSubmissionCommentDao.kt
@@ -13,15 +13,13 @@
  *     See the License for the specific language governing permissions and
  *     limitations under the License.
  */
-package com.instructure.student.room.entities.daos
+package com.instructure.pandautils.room.studentdb.entities.daos
 
 import androidx.lifecycle.LiveData
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.Query
-import com.instructure.canvasapi2.models.CanvasContext
-import com.instructure.student.room.entities.CreatePendingSubmissionCommentEntity
-import java.util.Date
+import com.instructure.pandautils.room.studentdb.entities.CreatePendingSubmissionCommentEntity
 
 @Dao
 interface CreatePendingSubmissionCommentDao {
